Move `/bigwins` to it's own file, and improve Discord embed
